dummy-rest-api/requirements.txt
astroid==2.0.4
certifi==2018.11.29
Django==3.0
django-cors-headers==3.5.0
django-filter==2.1.0
django-nested-admin==3.3.0
djangorestframework==3.12.1
djangorestframework-simplejwt==4.3.0
drf-extensions==0.6.0
isort==4.3.4
lazy-object-proxy==1.3.1
Markdown==3.0.1
mccabe==0.6.1
PyJWT==1.7.1
pylint==2.1.1
pytz==2018.9
wrapt==1.10.11
mysqlclient==1.4.2
Pillow==7.2.0